Propert y Even is a long and represents the number on each even row.

+++ PackItLib/src/Pack/Collation.cs
@@ -39,10 +39,10 @@ namespace PackIt.Pack
         /// <value> The total number of y coordinate. </value>
         public long CountY { get; set; }
 
-        /// <summary> Gets or sets the even. </summary>
+        /// <summary> Gets or sets the even row count. </summary>
         ///
-        /// <value> The even. </value>
-        public double Even { get; set; }
+        /// <value> The even row count. </value>
+        public long Even { get; set; }
 
         /// <summary> Gets or sets the nested. </summary>
         ///
